+++ src/main/java/me/ksyzov/accountmanager/Events.java
@@ -1,9 +1,9 @@
-package me.Vxrtrauter.accountmanager;
+package me.ksyzov.accountmanager;
 
-import me.Vxrtrauter.accountmanager.auth.Account;
-import me.Vxrtrauter.accountmanager.auth.SessionManager;
-import me.Vxrtrauter.accountmanager.gui.GuiAccountManager;
-import me.Vxrtrauter.accountmanager.utils.TextFormatting;
+import me.ksyzov.accountmanager.auth.Account;
+import me.ksyzov.accountmanager.auth.SessionManager;
+import me.ksyzov.accountmanager.gui.GuiAccountManager;
+import me.ksyzov.accountmanager.utils.TextFormatting;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiButton;
 import net.minecraft.client.gui.GuiDisconnected;
