+++ src/main/java/me/ksyzov/accountmanager/auth/AccountType.java
@@ -1,4 +1,4 @@
-package me.Vxrtrauter.accountmanager.auth;
+package me.ksyzov.accountmanager.auth;
 
 public enum AccountType {
     CRACKED, // For cracked (offline) accounts
