+++ src/main/java/me/ksyzov/accountmanager/gui/GuiAccountManager.java
@@ -1,13 +1,13 @@
-package me.Vxrtrauter.accountmanager.gui;
+package me.ksyzov.accountmanager.gui;
 
-import me.Vxrtrauter.accountmanager.AccountManager;
-import me.Vxrtrauter.accountmanager.auth.Account;
-import me.Vxrtrauter.accountmanager.auth.AccountType;
-import me.Vxrtrauter.accountmanager.auth.CrackedAuth;
-import me.Vxrtrauter.accountmanager.auth.MicrosoftAuth;
-import me.Vxrtrauter.accountmanager.auth.SessionManager;
-import me.Vxrtrauter.accountmanager.utils.Notification;
-import me.Vxrtrauter.accountmanager.utils.TextFormatting;
+import me.ksyzov.accountmanager.AccountManager;
+import me.ksyzov.accountmanager.auth.Account;
+import me.ksyzov.accountmanager.auth.AccountType;
+import me.ksyzov.accountmanager.auth.CrackedAuth;
+import me.ksyzov.accountmanager.auth.MicrosoftAuth;
+import me.ksyzov.accountmanager.auth.SessionManager;
+import me.ksyzov.accountmanager.utils.Notification;
+import me.ksyzov.accountmanager.utils.TextFormatting;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.*;
 import org.apache.commons.lang3.StringUtils;
