+++ src/main/java/me/ksyzov/accountmanager/utils/Notification.java
@@ -1,4 +1,4 @@
-package me.Vxrtrauter.accountmanager.utils;
+package me.ksyzov.accountmanager.utils;
 
 @SuppressWarnings("unused")
 public class Notification {
