+++ src/main/java/me/ksyzov/accountmanager/utils/TextFormatting.java
@@ -1,4 +1,4 @@
-package me.Vxrtrauter.accountmanager.utils;
+package me.ksyzov.accountmanager.utils;
 
 @SuppressWarnings("unused")
 public enum TextFormatting {
