+++ src/main/java/me/ksyzov/accountmanager/utils/UsernameGenerator.java
@@ -1,4 +1,4 @@
-package me.Vxrtrauter.accountmanager.utils;
+package me.ksyzov.accountmanager.utils;
 
 import javax.net.ssl.HttpsURLConnection;
 import java.io.BufferedReader;
